requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
cssselect2==0.7.0
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
numpy==1.23.4
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
redis==4.3.4
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
twilio==7.15.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
